+++ pkg/utils/utils.go
@@ -1,6 +1,11 @@
 package utils
 
-import "github.com/google/uuid"
+import (
+	"math/rand"
+	"time"
+
+	"github.com/google/uuid"
+)
 
 func Unique(items []string) []string {
 	keys := make(map[string]bool)
@@ -80,4 +85,23 @@ func DiffU(before, after []uuid.UUID) ([]uuid.UUID, []uuid.UUID, []uuid.UUID) {
 		}
 	}
 	return cu, uu, du
+}
+
+func GetRandomPassword(length int) string {
+	rand.Seed(time.Now().UnixNano())
+	digits := "0123456789"
+	specials := "~=+%^*/()[]{}/!@#$?|"
+	all := "ABCDEFGHIJKLMNOPQRSTUVWXYZ" +
+		"abcdefghijklmnopqrstuvwxyz" +
+		digits + specials
+	buf := make([]byte, length)
+	buf[0] = digits[rand.Intn(len(digits))]
+	buf[1] = specials[rand.Intn(len(specials))]
+	for i := 2; i < length; i++ {
+		buf[i] = all[rand.Intn(len(all))]
+	}
+	rand.Shuffle(len(buf), func(i, j int) {
+		buf[i], buf[j] = buf[j], buf[i]
+	})
+	return string(buf)
 }