+++ CHANGELOG.md
@@ -7,6 +7,12 @@ All notable changes to this project will be documented in this file.
 ## Fixed
 - Fix project id is recorded as part of cluster related auditlogs from [niravparikh05](https://github.com/niravparikh05)
 
+## Added
+- Enhance: Ability to set auto generated password during installation and force reset during first login [mabhi](https://github.com/mabhi)
+
+## Changed
+- Upgraded Ory Kratos to v0.10.1 [akshay196](https://github.com/akshay196)
+
 ## [0.1.9] - 2022-12-29
 
 ## Added